% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervals.R
\name{ComputeIntervals}
\alias{ComputeIntervals}
\title{Function for calculating intervals for suppressed tables.}
\usage{
ComputeIntervals(
  x,
  z,
  primary,
  suppressed,
  minVal = NULL,
  lpPackage = "lpSolve",
  gaussI = TRUE
)
}
\arguments{
\item{x}{ModelMatrix, as output from SSBtools::ModelMatrix}

\item{z}{numerical vector with length nrow(x). Corresponds to inner cell values}

\item{primary}{Vector indicating primary suppressed cells. Can be logical or
integer. If integer vector, indicates the columns of x which are considered
primary suppressed.}

\item{suppressed}{Vector indicating all suppressed cells. Can be logical or
integer. If integer vector, indicates the columns of x which are considered
suppressed.}

\item{minVal}{a known minimum value for table cells. Default NULL.}

\item{lpPackage}{The name of the package used to solve linear programs. Currently, only
'lpSolve' (default) and 'Rsymphony' are supported.}

\item{gaussI}{Boolean vector. If TRUE (default), GaussIndependent is used to
reduce size of linear program.}
}
\description{
This function solves linear programs to determine interval boundaries
for suppressed cells.
}
\details{
This function is still experimental.
}
\author{
Øyvind Langrsrud and Daniel Lupp
}
