% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressKDisclosure.R
\name{default_targeting}
\alias{default_targeting}
\title{Default \code{targeting} function for SuppressKDisclosure()}
\usage{
default_targeting(crossTable, identifying = NULL, disclosive = NULL, ...)
}
\arguments{
\item{crossTable}{A \code{crossTable}, possibly extended after applying
\code{mc_hierarchies}.}

\item{identifying}{A named list specifying identifying cells.}

\item{disclosive}{A named list specifying disclosive codes.}

\item{...}{Unused parameters.}
}
\value{
A named \code{targeting} list. See \code{\link[=SuppressKDisclosure]{SuppressKDisclosure()}}.
}
\description{
Generates a \code{targeting} specification for use with
\code{\link[=SuppressKDisclosure]{SuppressKDisclosure()}}. The function is actually used internally by
\code{\link[=KDisclosurePrimary]{KDisclosurePrimary()}}.
}
\details{
Since \code{\link[SSBtools:WildcardGlobbing]{SSBtools::WildcardGlobbing()}} is used, specifications may include
the wildcard characters \code{*} and \verb{?}, as well as the exclusion operator \code{!}.
}
\examples{

crossTable <- SSBtools::ModelMatrix(SSBtoolsData("example1"), 
     formula = ~age * eu + geo, crossTable = TRUE)$crossTable

default_targeting(crossTable)  # just NULL 

# geo identifying and age disclosive (age sensitive variable)
default_targeting(crossTable, identifying = list(age = "Total", geo = "*"), 
    disclosive = list(age = "*"))  # just NULL 


# Not ok to disclose 'EU' But ok to disclose 'Spain' with 'EU' known
default_targeting(crossTable, disclosive = list(geo = c("Portugal", "EU")))

}
