% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussSuppressDec.R
\name{GaussSuppressDec}
\alias{GaussSuppressDec}
\title{GaussSuppressDec}
\usage{
GaussSuppressDec(data, ..., output = "both", digits = 9, nRep = 1, rmse = pi/3)
}
\arguments{
\item{data}{Input daata as a data frame}

\item{...}{Further parameters to \code{\link{GaussSuppressionFromData}}}

\item{output}{output}

\item{digits}{digits}

\item{nRep}{nRep}

\item{rmse}{rmse}
}
\value{

}
\description{
GaussSuppressDec
}
\examples{
z1 <- SSBtoolsData("z1")
GaussSuppressDec(z1, 1:2, 3, nRep =3)$b
}
