% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDefault.R
\name{GetDefault}
\alias{GetDefault}
\title{Get default value of a function}
\usage{
GetDefault(fun, parameter, ifnotfound = NULL)
}
\arguments{
\item{fun}{A function}

\item{parameter}{parameter name}

\item{ifnotfound}{}
}
\value{
The default parameter, possibly evaluated.
}
\description{
The value may be found in a spec.
See \code{\link{PackageSpecs}}.
}
\details{
The result is evaluated if \code{\link{is.name}} returns \code{TRUE}.
}
\examples{
fun1 <- GetDefault(GaussSuppressionFromData, "candidates")
fun2 <- GetDefault(SuppressFewContributors, "primary")
fun3 <- GetDefault(SuppressDominantCells, "primary") 
}
\keyword{internal}
