% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressKDisclosure.R
\name{KDisclosurePrimary}
\alias{KDisclosurePrimary}
\title{Construct primary suppressed difference matrix}
\usage{
KDisclosurePrimary(
  data,
  x,
  crossTable,
  freqVar,
  mc_hierarchies = NULL,
  coalition = 1,
  upper_bound = Inf,
  identifying = NULL,
  disclosive = NULL,
  ...
)
}
\arguments{
\item{data}{a data.frame representing the data set}

\item{x}{ModelMatrix generated by parent function}

\item{crossTable}{crossTable generated by parent function}

\item{freqVar}{name of the frequency variable in \code{data}}

\item{mc_hierarchies}{a hierarchy representing meaningful combinations to be
protected. Default value is \code{NULL}.}

\item{coalition}{numeric vector of length one, representing possible size of an
attacking coalition. This parameter corresponds to the parameter k in the
definition of k-disclosure.}

\item{upper_bound}{numeric value representing minimum count considered safe.
Default set to \code{Inf}}

\item{identifying}{A data frame containing selected rows from \code{crossTable}.
The reference is to a \code{crossTable} data frame, possibly extended after applying
\code{mc_hierarchies}. Membership in the cells represented by these rows is regarded
as information that an intruder may already know. Alternatively, the parameter
can be specified as a function that generates this data frame. This works
similarly to supplied functions in \code{GaussSuppressionFromData()}. Note, however,
that the function operates on versions of \code{freq}, \code{x}, and \code{crossTable} that
reflect the use of \code{mc_hierarchies}, when applicable.}

\item{disclosive}{A data frame or a generating function that provides selected
rows from \code{crossTable}. If an intruder can infer membership in the cells
represented by these rows, this is considered an unacceptable disclosure.
This parameter works in a similar way to \code{identifying} described above.}

\item{...}{parameters passed to children functions}
}
\value{
dgCMatrix corresponding to primary suppressed cells
}
\description{
Function for constructing model matrix columns representing primary suppressed
difference cells
}
\author{
Daniel P. Lupp and Øyvind Langsrud
}
