% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedSuppression.R
\name{LinkedSuppression}
\alias{LinkedSuppression}
\title{Consistent Suppression of Linked Tables}
\usage{
LinkedSuppression(
  fun,
  data = NULL,
  ...,
  withinArg = NULL,
  linkedGauss,
  recordAware = TRUE,
  iterBackTracking = Inf,
  whenEmptyUnsuppressed = NULL
)
}
\arguments{
\item{fun}{A function: \code{\link{GaussSuppressionFromData}} or one of its wrappers such as
\code{\link{SuppressSmallCounts}} and \code{\link{SuppressDominantCells}}.}

\item{data}{The \code{data} argument to \code{fun}. When NULL \code{data} must be included in  \code{withinArg}.}

\item{...}{Arguments to \code{fun} that are not kept constant.}

\item{withinArg}{A list of named lists. Arguments to \code{fun} that are not kept constant.}

\item{linkedGauss}{See \link{parameter_linkedGauss}.}

\item{recordAware}{See \link{parameter_linkedGauss}.}

\item{iterBackTracking}{See \link{parameter_linkedGauss}.}

\item{whenEmptyUnsuppressed}{Parameter to \code{\link[SSBtools]{GaussSuppression}}.}
}
\value{
List of data frames
}
\description{
Consistent Suppression of Linked Tables
}
\examples{
 
#### Similar to parameter_linkedGauss example
# Trick "sector4 - sector4" and "geo - geo" to ensure same names in output 
output <- LinkedSuppression(data = SSBtoolsData("magnitude1"),
                 fun = SuppressDominantCells, 
                 withinArg = list(list(formula = ~(geo + eu) * sector2 + sector4 - sector4), 
                                  list(formula = ~eu:sector4 - 1 + geo - geo), 
                                  list(formula = ~geo + eu + sector4 - 1)), 
                 dominanceVar  = "value", 
                 pPercent = 10, 
                 contributorVar = "company",
                 singletonMethod = "none", 
                 linkedGauss = "consistent") 

####  Similar to LazyLinkedTables example:
z1 <- SSBtoolsData("z1")
z2 <- SSBtoolsData("z2")
z2b <- z2[3:5]  # As in ChainedSuppression example 
names(z2b)[1] <- "region" 
# As 'f' and 'e' in ChainedSuppression example. 
# 'A' 'annet'/'arbeid' suppressed in b[[1]], since suppressed in b[[3]].
b <- LinkedSuppression(fun = SuppressSmallCounts,
     linkedGauss = "back-tracking",  
     singletonMethod = "none",  # since For now singletonMethod must be none ... 
     recordAware = FALSE,
     withinArg = list(
       list(data = z1, dimVar = 1:2, freqVar = 3, maxN = 5), 
       list(data = z2b, dimVar = 1:2, freqVar = 3, maxN = 5), 
       list(data = z2, dimVar = 1:4, freqVar = 5, maxN = 1)))
}
