% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrimaryFromSuppressedData.R
\name{PrimaryFromSuppressedData}
\alias{PrimaryFromSuppressedData}
\alias{ForcedFromSuppressedData}
\alias{NotPrimaryFromSuppressedData}
\title{\code{primary} and \code{forced} from suppressed data}
\usage{
PrimaryFromSuppressedData(
  x,
  crossTable,
  suppressedData,
  forcedData = FALSE,
  totCode = FindTotCode2(x, crossTable),
  ...
)

ForcedFromSuppressedData(..., forcedData = TRUE)

NotPrimaryFromSuppressedData(..., forcedData = TRUE)
}
\arguments{
\item{x}{A (sparse) dummy matrix}

\item{crossTable}{crossTable generated by parent function}

\item{suppressedData}{A data frame or a list of data frames as output from \code{\link{GaussSuppressionFromData}}. I variable \code{suppressed} is not included. All rows are considered suppressed.}

\item{forcedData}{When \code{TRUE}, the suppressed coding is swapped.}

\item{totCode}{A named list of totals codes}
}
\value{
Logical vector to be used as \code{\link{GaussSuppression}} input
}
\description{
Function for \code{\link{GaussSuppressionFromData}}
}
\details{
\code{ForcedFromSuppressedData}  uses \code{forcedData = TRUE} and hence a  vector to be use as forced is generated.
\code{NotPrimaryFromSuppressedData} is similar, but \code{TRUE} elements are replaced by \code{NA}'s.
Hence the result can be used as an extra primary vector to ensure that code combinations
not suppressed according to \code{suppressedData} are forced not to be primary suppressed.
}
