% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeLimits.R
\name{RangeOutputFunction}
\alias{RangeOutputFunction}
\title{Function used in RangeLimitsDefault example}
\usage{
RangeOutputFunction(...)
}
\arguments{
\item{...}{parameters}
}
\value{
Aggregated data with suppression information
}
\description{
Function used in RangeLimitsDefault example
}
\keyword{internal}
