% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussSuppressionFromData.R
\name{SingletonDefault}
\alias{SingletonDefault}
\title{SingletonDefault}
\usage{
SingletonDefault(data, freqVar, protectZeros, secondaryZeros, ...)
}
\arguments{
\item{data}{data}

\item{freqVar}{freqVar}

\item{protectZeros}{protectZeros}

\item{secondaryZeros}{secondaryZeros}

\item{...}{Unused parameters}
}
\value{
singleton, \code{\link{GaussSuppression}} input
}
\description{
Function for \code{\link{GaussSuppressionFromData}}
}
\keyword{internal}
