% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressKDisclosure.R
\name{SuppressKDisclosure}
\alias{SuppressKDisclosure}
\title{K-disclosure suppression}
\usage{
SuppressKDisclosure(
  data,
  coalition = 0,
  mc_hierarchies = NULL,
  upper_bound = Inf,
  dimVar = NULL,
  formula = NULL,
  hierarchies = NULL,
  freqVar = NULL,
  targeting = NULL,
  print_frames = FALSE,
  ...,
  spec = PackageSpecs("kDisclosureSpec")
)
}
\arguments{
\item{data}{a data.frame representing the data set}

\item{coalition}{numeric vector of length one, representing possible size of an
attacking coalition. This parameter corresponds to the parameter k in the
definition of k-disclosure.}

\item{mc_hierarchies}{a hierarchy representing meaningful combinations to be
protected. Default value is \code{NULL}.}

\item{upper_bound}{numeric value representing minimum count considered safe.
Default set to \code{Inf}}

\item{dimVar}{The main dimensional variables and additional aggregating
variables. This parameter can be  useful when hierarchies and formula are
unspecified.}

\item{formula}{A model formula}

\item{hierarchies}{List of hierarchies, which can be converted by
\code{\link[SSBtools]{AutoHierarchies}}. Thus, the variables can also be coded by
\code{"rowFactor"} or \code{""}, which correspond to using the categories in the data.}

\item{freqVar}{name of the frequency variable in \code{data}}

\item{targeting}{NULL, a list, or a function that returns a list specifying
attribute disclosure scenarios. See Details.}

\item{print_frames}{Logical. If TRUE, additional data frames are printed to
the console. When \code{mc_hierarchies} is used, this includes a
data frame with hidden results. In addition, a data frame containing the
primary suppressed difference cells is printed. The default is FALSE.}

\item{...}{parameters passed to children functions}

\item{spec}{\code{NULL} or a named list of arguments that will act as default values.}
}
\value{
A data.frame containing the publishable data set, with a boolean
variable \verb{$suppressed} representing cell suppressions.
}
\description{
A function for suppressing frequency tables using the k-disclosure method.
}
\details{
The \code{targeting} specification is a named list that may contain the following
optional elements. References to \code{crossTable} below refer to a data frame
that may be extended after applying \code{mc_hierarchies}.

\describe{
\item{\code{identifying}}{A data frame containing selected rows from
\code{crossTable}. Membership in the cells represented by these rows is regarded
as information that an intruder may already know. If omitted, it defaults
to \code{crossTable}.}

\item{\code{disclosive}}{A data frame containing selected rows from
\code{crossTable}. If an intruder can infer membership in the cells represented
by these rows, this is considered an unacceptable disclosure, subject to
any further specification provided by \code{is_disclosive}. If omitted, it
defaults to \code{crossTable}.}

\item{\code{is_disclosive}}{A data frame with the same structure as
\code{disclosive}, but with logical variables. It indicates which codes in
\code{disclosive} are regarded as disclosive. When specified, disclosure is
assessed by which codes within a revealed cell are disclosed. If omitted,
it is equivalent to a data frame where all elements are \code{TRUE}.}
}

The argument \code{targeting} may also be a function that returns such a list.
This works similarly to supplied functions in \code{GaussSuppressionFromData()}.
Note, however, that the function operates on possibly extended versions of
\code{freq}, \code{x}, and \code{crossTable} that reflect the use of \code{mc_hierarchies}, when
applicable.
}
\examples{
# data
data <- SSBtools::SSBtoolsData("mun_accidents")

# hierarchies as DimLists
mun <- data.frame(levels = c("@", rep("@@", 6)),
codes = c("Total", paste("k", 1:6, sep = "")))
inj <- data.frame(levels = c("@", "@@" ,"@@", "@@", "@@"),
codes = c("Total", "serious", "light", "none", "unknown"))
dimlists <- list(mun = mun, inj = inj)

inj2 <- data.frame(levels = c("@", "@@", "@@@" ,"@@@", "@@", "@@"),
codes = c("Total", "injured", "serious", "light", "none", "unknown"))
inj3 <- data.frame(levels = c("@", "@@", "@@" ,"@@", "@@"),
codes = c( "shadowtotal", "serious", "light", "none", "unknown"))
mc_dimlist <- list(inj = inj2)
mc_nomargs <- list(inj = inj3)

#' # Example with formula, no meaningful combination
out <- SuppressKDisclosure(data, coalition = 1, freqVar = "freq", formula = ~mun*inj)

# Example with hierarchy and meaningful combination
out2 <- SuppressKDisclosure(data, coalition = 1, freqVar = "freq",
hierarchies = dimlists, mc_hierarchies = mc_dimlist)

#' # Example of table without mariginals, and mc_hierarchies to protect
out3 <- SuppressKDisclosure(data, coalition = 1, freqVar = "freq",
formula = ~mun:inj, mc_hierarchies = mc_nomargs )
}
\author{
Daniel P. Lupp and Øyvind Langsrud
}
