% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressSmallCounts.R
\name{SuppressSmallCounts}
\alias{SuppressSmallCounts}
\title{Small count frequency table suppression.}
\usage{
SuppressSmallCounts(
  data,
  maxN,
  freqVar = NULL,
  dimVar = NULL,
  hierarchies = NULL,
  formula = NULL,
  ...,
  spec = PackageSpecs("smallCountSpec")
)
}
\arguments{
\item{data}{Input data as a data frame}

\item{maxN}{Suppression parameter. Cells with frequency \verb{<= maxN} are set as primary suppressed.
Using the default \code{primary} function, \code{maxN} is by default set to \code{3}. See details.}

\item{freqVar}{A single variable holding counts (name or number).}

\item{dimVar}{The main dimensional variables and additional aggregating variables. This parameter can be  useful when hierarchies and formula are unspecified.}

\item{hierarchies}{List of hierarchies, which can be converted by \code{\link{AutoHierarchies}}.
Thus, the variables can also be coded by \code{"rowFactor"} or \code{""}, which correspond to using the categories in the data.}

\item{formula}{A model formula}

\item{...}{Further arguments to be passed to the supplied functions and to \code{\link{ModelMatrix}} (such as \code{inputInOutput} and \code{removeEmpty}).}

\item{spec}{\code{NULL} or a named list of arguments that will act as default values.}
}
\value{
data frame containing aggregated data and suppression information.
}
\description{
This is a wrapper function of \code{\link{GaussSuppressionFromData}}
for small count frequency suppression. For common applications, the \code{spec}
parameter can be adjusted, see \code{\link{PackageSpecs}} for more
information. See Details for more information on function call customization.
}
\details{
The specs provided in the package (see \code{\link{PackageSpecs}})
provide common parameter setups for small count suppression. However, it might
be necessary to customize the parameters further. In this case, certain
parameters from \code{\link{GaussSuppressionFromData}}
might need adjusting from the values provided by the package specs. In
particular, the parameters \code{protectZeros} (should zeros be primary
suppressed), \code{extend0} (should empty cells be added before primary
suppression), and \code{secondaryZeros} (should zero frequency cells be
candidates for secondary suppression) might be of interest. The examples
below illustrate how to override parameters specified by a spec. Note
that this is only possible if \code{specLock = FALSE}.
}
\examples{
mun <- c("k1", "k2", "k3", "k4", "k5", "k6")
inj <- c("serious", "light", "none", "unknown")
data <- expand.grid(mun, inj)
names(data) <- c("mun", "inj")
data$freq <- c(4,5,3,4,1,6,
0,0,2,1,0,0,
0,1,1,4,0,0,
0,0,0,0,0,0)

SuppressSmallCounts(data = data, maxN = 3, dimVar = 1:2, freqVar = 3)
# override default spec
SuppressSmallCounts(data = data, maxN = 3, dimVar = 1:2, freqVar = 3, protectZeros = FALSE)
}
