% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputFunction.R
\name{ToyOutputFunction}
\alias{ToyOutputFunction}
\title{Toy output function}
\usage{
ToyOutputFunction(..., words = c("Yes", "No"))
}
\arguments{
\item{...}{}

\item{words}{}
}
\description{
Example of using special temporary feature.
Possible development function as input.
}
\examples{
 SuppressDominantCells(data = SSBtoolsData("magnitude1"), 
                       numVar = "value", 
                       formula = ~sector2 * geo + sector4 * eu, 
                       contributorVar = "company", 
                       n = 1:2, k = c(80, 99), 
                       output = ToyOutputFunction, 
                       words = rownames(mtcars))
}
\keyword{internal}
