% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_from_mc.R
\name{X_from_mc}
\alias{X_from_mc}
\alias{X_from_mc_remove}
\alias{X_from_mc_noinner}
\alias{X_from_mc_remove_noinner}
\title{Function to be used within \code{KDisclosurePrimary}}
\usage{
X_from_mc(
  x,
  crossTable,
  mc_hierarchies,
  removeIncomplete = FALSE,
  returnNewCrossTable = FALSE,
  noInner = FALSE,
  ...
)

X_from_mc_remove(..., removeIncomplete = TRUE)

X_from_mc_noinner(..., noInner = TRUE)

X_from_mc_remove_noinner(..., removeIncomplete = TRUE, noInner = TRUE)
}
\arguments{
\item{x}{ModelMatrix generated by parent function}

\item{crossTable}{crossTable generated by parent function}

\item{mc_hierarchies}{a hierarchy representing meaningful combinations to be
protected. Default value is \code{NULL}.}

\item{removeIncomplete}{When \code{TRUE},  the input code contributions are checked and incomplete entries are removed.}

\item{returnNewCrossTable}{When \code{TRUE}, the crossTable corresponding to the created x-matrix is also returned (in a list)}

\item{noInner}{When \code{TRUE},  more efficient generation of model matrix.  (removing inner cells according to colSums not needed)}

\item{...}{parameters passed to children functions}
}
\description{
Function to be used within \code{KDisclosurePrimary}
}
\author{
Øyvind Langsrud
}
\keyword{internal}
