% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussSuppressionFromData.R
\name{CandidatesDefault}
\alias{CandidatesDefault}
\title{CandidatesDefault}
\usage{
CandidatesDefault(freq, x, secondaryZeros = FALSE, weight, ...)
}
\arguments{
\item{freq}{freq}

\item{x}{x}

\item{secondaryZeros}{secondaryZeros}

\item{...}{Unused parameters}
}
\value{
candidates, \code{\link{GaussSuppression}} input
}
\description{
Function for \code{\link{GaussSuppressionFromData}}
}
\keyword{internal}
