% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressKDisclosure.R
\name{default_targeting}
\alias{default_targeting}
\title{Default \code{targeting} function for SuppressKDisclosure()}
\usage{
default_targeting(crossTable, x, identifying = NULL, disclosive = NULL, ...)
}
\arguments{
\item{crossTable}{A \code{crossTable}, possibly extended after applying
\code{mc_hierarchies}.}

\item{x}{The model matrix, \code{x}, possibly extended after applying
\code{mc_hierarchies}.}

\item{identifying}{A named list specifying identifying cells.}

\item{disclosive}{A named list specifying disclosive codes.}

\item{...}{Unused parameters.}
}
\value{
A named \code{targeting} list. See \code{\link[=SuppressKDisclosure]{SuppressKDisclosure()}}.
}
\description{
Generates a \code{targeting} specification for use with
\code{\link[=SuppressKDisclosure]{SuppressKDisclosure()}}. The function is actually used internally by
\code{\link[=KDisclosurePrimary]{KDisclosurePrimary()}}.
}
\details{
Since \code{\link[SSBtools:WildcardGlobbing]{SSBtools::WildcardGlobbing()}} is used, specifications may include
the wildcard characters \code{*} and \verb{?}, as well as the exclusion operator \code{!}.
}
\examples{

mm <- SSBtools::ModelMatrix(SSBtoolsData("example1"), 
     formula = ~age * eu + geo, crossTable = TRUE)
crossTable <- mm$crossTable
x <- mm$modelMatrix      

default_targeting(crossTable, x)  # just NULL 

# geo identifying and age disclosive (age sensitive variable)
a2 <- default_targeting(crossTable, x, 
                        identifying = "geo", 
                        disclosive = "age")
a1 <- default_targeting(crossTable, x, 
                        identifying = list(age = "Total", geo = "*"), 
                        disclosive = list(age = "*")) 
identical(a1, a2)
a1                         
                  
                  
# Not ok to disclose 'EU' and 'Portugal'
# But ok to disclose 'Spain' with 'EU' known
# and also ok to disclose 'Spain' in other table cells without 'EU' as marginal  
default_targeting(crossTable, x, 
                  disclosive = list(geo = c("Portugal", "EU")))
                  
# As above but now also ok to disclose 'Portugal' from 'EU' known,
# since protection only considers 'age' identifying.                   
default_targeting(crossTable, x, 
                  identifying = "age",
                  disclosive = list(geo = c("Portugal", "EU")))                 

}
