% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussSuppressDec.R
\name{GaussSuppressDec}
\alias{GaussSuppressDec}
\title{GaussSuppressDec}
\usage{
GaussSuppressDec(
  data,
  ...,
  output = NULL,
  digits = 9,
  nRep = NULL,
  rmse = pi/3,
  sparseLimit = 500,
  rndSeed = 123
)
}
\arguments{
\item{data}{Input daata as a data frame}

\item{...}{Further parameters to \code{\link{GaussSuppressionFromData}}}

\item{output}{output}

\item{digits}{digits}

\item{nRep}{nRep}

\item{rmse}{rmse}

\item{rndSeed}{If non-NULL, a random generator seed to be used locally within the function without affecting the random value stream in R.}
}
\value{

}
\description{
GaussSuppressDec
}
\examples{
z1 <- SSBtoolsData("z1")
GaussSuppressDec(z1, 1:2, 3)
GaussSuppressDec(z1, freqVar = "ant", formula = ~ region + hovedint, maxN = 10)
}
