% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primary.R
\name{PrimaryDefault}
\alias{PrimaryDefault}
\title{Default primary function}
\usage{
PrimaryDefault(freq, maxN = 3, protectZeros = TRUE, ...)
}
\arguments{
\item{freq}{Vector of output frequencies}

\item{maxN}{Cells with frequency \verb{<= maxN} are set as primary suppressed.}

\item{protectZeros}{When \code{TRUE}, cells with zero frequency are set as primary suppressed.}

\item{...}{Unused parameters}
}
\value{
primary, \code{\link[SSBtools]{GaussSuppression}} input
}
\description{
Function for \code{\link{GaussSuppressionFromData}}
}
