% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussSuppressionFromData.R
\name{PrimaryDefault}
\alias{PrimaryDefault}
\title{PrimaryDefault}
\usage{
PrimaryDefault(freq, x, maxN = 3, protectZeros = TRUE, ...)
}
\arguments{
\item{freq}{freq}

\item{x}{x}

\item{maxN}{maxN}

\item{protectZeros}{protectZeros}

\item{...}{Unused parameters}
}
\value{
primary, \code{\link{GaussSuppression}} input
}
\description{
Function for \code{\link{GaussSuppressionFromData}}
}
\keyword{internal}
