% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressDirectDisclosure.R
\name{SuppressDirectDisclosure}
\alias{SuppressDirectDisclosure}
\title{Suppression of directly-disclosive cells}
\usage{
SuppressDirectDisclosure(
  data,
  dimVar,
  freqVar,
  secondaryZeros = T,
  singleton = NULL,
  ...
)
}
\arguments{
\item{data}{the input data}

\item{dimVar}{character vector containing variable names for the output table}

\item{freqVar}{variable name containing frequency counts}

\item{...}{optional parameters that can be passed to the primary suppression
method. See \link[SSBtools]{FindDisclosiveCells} for details.}
}
\value{
data.frame containing the result of the suppression
}
\description{
Function for suppressing disclosive cells in frequency tables. The method
detects and primary suppresses directly-disclosive cells with the
\link[SSBtools]{FindDisclosiveCells} function, and applies a secondary suppression
using Gauss suppression (see \link{GaussSuppressionFromData}).
}
\details{
Currently, the method has no support for hierarchical data.
}
\examples{
tex <- data.frame(v1 = rep(c('a', 'b', 'c'), times = 4),
                  v2 = c('i','i', 'i','h','h','h','i','i','i','h','h','h'),
                  v3 = c('y', 'y', 'y', 'y', 'y', 'y','z','z', 'z', 'z', 'z', 'z'),
                  freq = c(0,0,5,0,2,3,1,0,3,1,1,2))
SuppressDirectDisclosure(tex, c("v1", "v2", "v3"), "freq")
SuppressDirectDisclosure(tex, c("v1", "v2", "v3"), "freq", coalition = 2, unknown.threshold = 10)
}
