% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressDominantCells.R
\name{SuppressDominantCells}
\alias{SuppressDominantCells}
\title{Suppress volume tables using dominant cell primary suppression.}
\usage{
SuppressDominantCells(
  data,
  n,
  k,
  numVar = NULL,
  dimVar = NULL,
  hierarchies = NULL,
  formula = NULL,
  preAggVar = NULL,
  sWeightVar = NULL,
  ...,
  spec = PackageSpecs("dominanceSpec")
)
}
\arguments{
\item{data}{Input data as a data frame}

\item{n}{parameter \code{n} in dominance rule.}

\item{k}{parameter \code{k} in dominance rule.}

\item{numVar}{numerical variable to be aggregated and used in dominance rule}

\item{dimVar}{The main dimensional variables and additional aggregating variables. This parameter can be  useful when hierarchies and formula are unspecified.}

\item{hierarchies}{List of hierarchies, which can be converted by \code{\link{AutoHierarchies}}.
Thus, the variables can also be coded by \code{"rowFactor"} or \code{""}, which correspond to using the categories in the data.}

\item{formula}{A model formula}

\item{preAggVar}{Extra variables to be used as grouping elements in the dominance rule.}

\item{sWeightVar}{Name of variable which represents sampling weights to be used
in dominance rule}

\item{...}{Further arguments to be passed to the supplied functions and to \code{\link{ModelMatrix}} (such as \code{inputInOutput} and \code{removeEmpty}).}

\item{spec}{\code{NULL} or a named list of arguments that will act as default values.}
}
\value{
data frame containing aggregated data and suppression information.
}
\description{
Suppress volume tables using dominant cell primary suppression.
}
