% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressKDisclosure.R
\name{SuppressKDisclosure}
\alias{SuppressKDisclosure}
\title{K-disclosure suppression}
\usage{
SuppressKDisclosure(
  data,
  coalition = 0,
  mc_hierarchies = NULL,
  upper_bound = Inf,
  dimVar = NULL,
  formula = NULL,
  hierarchies = NULL,
  freqVar = NULL,
  targeting = default_targeting,
  identifying = NULL,
  disclosive = NULL,
  print_frames = FALSE,
  ...,
  spec = PackageSpecs("kDisclosureSpec")
)
}
\arguments{
\item{data}{a data.frame representing the data set}

\item{coalition}{numeric vector of length one, representing possible size of an
attacking coalition. This parameter corresponds to the parameter k in the
definition of k-disclosure.}

\item{mc_hierarchies}{a hierarchy representing meaningful combinations to be
protected. Default value is \code{NULL}.}

\item{upper_bound}{numeric value representing minimum count considered safe.
Default set to \code{Inf}}

\item{dimVar}{The main dimensional variables and additional aggregating
variables. This parameter can be  useful when hierarchies and formula are
unspecified.}

\item{formula}{A model formula}

\item{hierarchies}{List of hierarchies, which can be converted by
\code{\link[SSBtools]{AutoHierarchies}}. Thus, the variables can also be coded by
\code{"rowFactor"} or \code{""}, which correspond to using the categories in the data.}

\item{freqVar}{name of the frequency variable in \code{data}}

\item{targeting}{The mechanism underlying the interpretation of
\code{identifying} and \code{disclosive}. See Details in \code{\link[=KDisclosurePrimary]{KDisclosurePrimary()}}.}

\item{identifying}{Specification of information that an intruder may already
know. The specification is subject to the same requirements as \code{disclosive}
below. If not all variables are included, total codes for the missing
variables are derived automatically. This requires that the overall total
is included as an output row.}

\item{disclosive}{Specification of information considered unacceptable to
disclose. Either a character vector of variable names, or a named list with
variable names as names and specified codes as values. The wildcard
characters \code{*} and \verb{?}, as well as the exclusion operator \code{!}, may be used,
since \code{\link[SSBtools:WildcardGlobbing]{SSBtools::WildcardGlobbing()}} is applied.}

\item{print_frames}{Logical. If TRUE, additional data frames are printed to
the console. When \code{mc_hierarchies} is used, this includes a
data frame with hidden results. In addition, a data frame containing the
primary suppressed difference cells is printed. The default is FALSE.}

\item{...}{parameters passed to children functions}

\item{spec}{\code{NULL} or a named list of arguments that will act as default values.}
}
\value{
A data.frame containing the publishable data set, with a boolean
variable \verb{$suppressed} representing cell suppressions.
}
\description{
A function for suppressing frequency tables using the k-disclosure method.
}
\details{
The argument \code{targeting} may also be a function that returns such a list.
This works similarly to supplied functions in \code{GaussSuppressionFromData()}.
Note, however, that the function operates on possibly extended versions of
\code{freq}, \code{x}, and \code{crossTable} that reflect the use of \code{mc_hierarchies}, when
applicable.
}
\examples{
# data
mun_a <- SSBtools::SSBtoolsData("mun_accidents")

# Function to print output in wide format, marking suppressed values with `*`
show_out <- function(out) {  
  out$freq = sprintf("\%s\%s", out$freq, c(" ","*")[1+out$suppressed])
  a <- reshape(out[1:3], idvar = "mun", timevar = "inj", direction = "wide", )
  names(a) <- sub("^freq\\\\.", "", names(a))
  print(a)}

# hierarchies as DimLists
mun <- data.frame(levels = c("@", rep("@@", 6)),
codes = c("Total", paste("k", 1:6, sep = "")))
inj <- data.frame(levels = c("@", "@@" ,"@@", "@@", "@@"),
codes = c("Total", "serious", "light", "none", "unknown"))
dimlists <- list(mun = mun, inj = inj)

inj2 <- data.frame(levels = c("@", "@@", "@@@" ,"@@@", "@@", "@@"),
codes = c("Total", "injured", "serious", "light", "none", "unknown"))
inj3 <- data.frame(levels = c("@", "@@", "@@" ,"@@", "@@"),
codes = c( "shadowtotal", "serious", "light", "none", "unknown"))
mc_dimlist <- list(inj = inj2)
mc_nomargs <- list(inj = inj3)

#' # Example with formula, no meaningful combination
out <- SuppressKDisclosure(mun_a, coalition = 1, freqVar = "freq", formula = ~mun*inj)
show_out(out)

# Example with hierarchy and meaningful combination
out2 <- SuppressKDisclosure(mun_a, coalition = 1, freqVar = "freq",
                       hierarchies = dimlists, mc_hierarchies = mc_dimlist)
show_out(out2)

#' # Example of table without mariginals, and mc_hierarchies to protect
out3 <- SuppressKDisclosure(mun_a, coalition = 1, freqVar = "freq",
                       formula = ~mun:inj, mc_hierarchies = mc_nomargs )
show_out(out3)


### Examples with identifying and disclosive ###

mun_b <- SSBtools::SSBtoolsData("mun_accidents")
mun_b$freq <- c(0,5,3,4,1,0,
                0,0,2,0,0,6,
                4,1,0,4,0,0,
                0,0,0,0,0,0)
                
out_d <- SuppressKDisclosure(mun_b, coalition = 1, freqVar = "freq",
                                 formula = ~mun*inj, disclosive= "inj")
show_out(out_d)                                                    
                

out_d1 <- SuppressKDisclosure(mun_b, coalition = 1, freqVar = "freq",
                              formula = ~mun*inj, mc_hierarchies = mc_dimlist,
                              disclosive = list(mun =  "k3", inj = "injured"))
show_out(out_d1)                             

out_d2 <- SuppressKDisclosure(mun_b, coalition = 1, freqVar = "freq",
                              formula = ~mun*inj, 
                              disclosive = list(inj = "serious", mun = "k3"))
show_out(out_d2)                         

out_i1 <- SuppressKDisclosure(mun_b, coalition = 1, freqVar = "freq",
                              formula = ~mun*inj, identifying = "mun")
show_out(out_i1)                            
 
out_i2 <- SuppressKDisclosure(mun_b, coalition = 1, freqVar = "freq",
                              formula = ~mun*inj, identifying = "inj")
show_out(out_i2)


# Same example as out_d, but with cells forced to be published, yielding unsafe table
out_unsafe <- SuppressKDisclosure(mun_b, coalition = 1, freqVar = "freq",
                                 formula = ~mun*inj, disclosive = "inj", 
                                 forced = c(12,14,15), output = "all")
show_out(out_unsafe$publish)

# colnames in $unsafe give an indication as to which cells/differences are unsafe
colnames(out_unsafe$unsafe)
                                       
}
\author{
Daniel P. Lupp and Øyvind Langsrud
}
