% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_from_mc.R
\name{X_from_mc}
\alias{X_from_mc}
\title{Possible \code{mc_function} to be used within \code{KDisclosurePrimary}}
\usage{
X_from_mc(
  x,
  crossTable,
  mc_hierarchies,
  removeIncomplete = FALSE,
  returnNewCrossTable = FALSE,
  ...
)
}
\arguments{
\item{x}{ModelMatrix generated by parent function}

\item{crossTable}{crossTable generated by parent function}

\item{mc_hierarchies}{a hierarchy representing meaningful combinations to be
protected}

\item{removeIncomplete}{When \code{TRUE},  the input code contributions are checked and incomplete entries are removed.}

\item{returnNewCrossTable}{When \code{TRUE}, the crossTable corresponding to the created x-matrix is also returned (in a list)}

\item{...}{parameters passed to children functions}
}
\description{
Possible \code{mc_function} to be used within \code{KDisclosurePrimary}
}
\keyword{internal}
